Fix test failures due to change in stringification
In Swift 3, `String(describing: ...)` formats `["a": "b"]` as such,
however formats `"a"` as simply `a`, i.e. without quotes. This leads
to discrepancies when Nimble stringifies actual value in the form of a
`Dictionary` (i.e. leaf `String`s have quotes), and when JSONMatcher
stringifies the expected value in the form of
DictionaryElement (i.e. leaf `String`s, in this case `StringElement`
having no quotes). The solution to ensure consitency between the two
is therefore to stringify `StringElement`s with quotes around the
actual string.